dev-requirements.txt
beautifulsoup4==4.9.1
behave==1.2.6
behaving==1.5.6
ckanapi==4.3
flake8==3.8.3
mock
pyfakefs==2.7
pytest-ckan
pytest-cov
selenium==4.8.2
splinter>=0.13.0
faker==18.3.0
